Reorganized App.jsx into multiple files
