--- frontend/src/Header.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Header = () => {
-  return (
-    <header style={{ padding: '20px', backgroundColor: '#333', color: 'white' }}>
-      <h1>TripSync</h1>
-    </header>
-  );
-};
-
-export default Header;
