+++ frontend/src/components/Icons.jsx
@@ -1,0 +1,40 @@
+import React from 'react';
+
+// SVG Icons
+const CalendarIcon = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" className="h-10 w-10" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z" />
+  </svg>
+);
+
+const VoteIcon = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" className="h-10 w-10" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />
+  </svg>
+);
+
+const MoneyIcon = () => (
+    <svg xmlns="http://www.w3.org/2000/svg" className="h-10 w-10" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 6v12m-3-2.818l.879.659c1.171.879 3.07.879 4.242 0 1.172-.879 1.172-2.303 0-3.182C13.536 12.219 12.768 12 12 12c-.725 0-1.45-.22-2.003-.659-1.106-.879-1.106-2.303 0-3.182s2.9-.879 4.006 0l.415.33M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+    </svg>
+);
+
+const BasecampIcon = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" className="h-8 w-8" viewBox="0 0 24 24" fill="currentColor" stroke="currentColor" strokeWidth="1">
+    <path d="M23.869 17.31 15.813 5.864c-0.003 -0.004 -0.006 -0.007 -0.009 -0.011a0.708 0.708 0 0 0 -0.095 -0.106q-0.009 -0.008 -0.018 -0.016a0.766 0.766 0 0 0 -0.059 -0.045l-0.003 -0.002a0.67 0.67 0 0 0 -0.069 -0.041c-0.005 -0.002 -0.01 -0.004 -0.014 -0.007a0.67 0.67 0 0 0 -0.058 -0.026q-0.01 -0.004 -0.021 -0.008a0.67 0.67 0 0 0 -0.056 -0.017q-0.011 -0.003 -0.023 -0.006a0.766 0.766 0 0 0 -0.059 -0.011c-0.007 -0.001 -0.014 -0.002 -0.021 -0.003a0.766 0.766 0 0 0 -0.081 -0.005h-6.453a0.718 0.718 0 0 0 -0.587 0.305L0.131 17.31A0.718 0.718 0 0 0 0.718 18.441h22.564a0.718 0.718 0 0 0 0.587 -1.131m-9.361 -0.305h-5.955l5.955 -8.461zM9.146 6.995h4.697L6.798 17.005H2.101zm6.798 10.01V8.544l5.955 8.461z"/>
+  </svg>
+);
+
+const LightIcon = () => (
+  <svg className="swap-on fill-current w-8 h-8" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
+    <path d="M5.64,17l-.71.71a1,1,0,0,0,0,1.41,1,1,0,0,0,1.41,0l.71-.71A1,1,0,0,0,5.64,17ZM5,12a1,1,0,0,0-1-1H3a1,1,0,0,0,0,2H4A1,1,0,0,0,5,12Zm7-7a1,1,0,0,0,1-1V3a1,1,0,0,0-2,0V4A1,1,0,0,0,12,5ZM5.64,7.05a1,1,0,0,0,.7.29,1,1,0,0,0,.71-.29,1,1,0,0,0,0-1.41l-.71-.71A1,1,0,0,0,4.93,6.34Zm12,.29a1,1,0,0,0,.7-.29l.71-.71a1,1,0,1,0-1.41-1.41L17,5.64a1,1,0,0,0,0,1.41A1,1,0,0,0,17.66,7.34ZM21,11H20a1,1,0,0,0,0,2h1a1,1,0,0,0,0-2Zm-9,8a1,1,0,0,0-1,1v1a1,1,0,0,0,2,0V20A1,1,0,0,0,12,19ZM18.36,17A1,1,0,0,0,17,18.36l.71.71a1,1,0,0,0,1.41,0,1,1,0,0,0,0-1.41ZM12,6.5A5.5,5.5,0,1,0,17.5,12,5.51,5.51,0,0,0,12,6.5Zm0,9A3.5,3.5,0,1,1,15.5,12,3.5,3.5,0,0,1,12,15.5Z" />
+  </svg>
+);
+
+const DarkIcon = () => (
+  <svg className="swap-off fill-current w-8 h-8" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
+    <path d="M21.64,13a1,1,0,0,0-1.05-.14,8.05,8.05,0,0,1-3.37.73A8.15,8.15,0,0,1,9.08,5.49a8.59,8.59,0,0,1,.25-2A1,1,0,0,0,8,2.36,10.14,10.14,0,1,0,22,14.05,1,1,0,0,0,21.64,13Zm-9.5,6.69A8.14,8.14,0,0,1,7.08,5.22v.27A10.15,10.15,0,0,0,17.22,15.63a9.79,9.79,0,0,0,2.1-.22A8.11,8.11,0,0,1,12.14,19.73Z" />
+  </svg>
+);
+
+export { CalendarIcon, VoteIcon, MoneyIcon, BasecampIcon, LightIcon, DarkIcon };
